Add script to install brew and several common tools using brew
Eventually I picture using a script that would read a list of applications
to install and ensure they're installed/up to date. Perhaps even be able
to offer optional installation for the case of setting up a work vs. home
machine. For now I'm just keeping it simple with a shell script though.